Add basic gameplay functionality
Pretty much playable, if awkward. There is no indication each
turn whether the player was right or wrong, and in strict
mode player is just dumped back to a new game with no animation or
wait time. Also still need to complete all the button presses for a
turn before being compared against Simon's play.

On branch master
Changes to be committed: